+++ src/endpoint.rs
@@ -1,16 +1,16 @@
 use crate::endpoint_memory::{EndpointBuffer, EndpointBufferState};
 use crate::ral::{endpoint0_out, endpoint_in, endpoint_out, modify_reg, read_reg, write_reg};
-use crate::target::interrupt::{self, CriticalSection, Mutex};
 use crate::target::{fifo_write, UsbRegisters};
 use crate::transition::EndpointDescriptor;
 use crate::UsbPeripheral;
 use core::cell::RefCell;
 use core::ops::{Deref, DerefMut};
+use critical_section::{CriticalSection, Mutex};
 use usb_device::endpoint::EndpointAddress;
 use usb_device::{Result, UsbDirection, UsbError};
 
 pub fn set_stalled(usb: UsbRegisters, address: EndpointAddress, stalled: bool) {
-    interrupt::free(|_| match address.direction() {
+    critical_section::with(|_| match address.direction() {
         UsbDirection::Out => {
             let ep = usb.endpoint_out(address.index() as usize);
             modify_reg!(endpoint_out, ep, DOEPCTL, STALL: stalled as u32);
@@ -71,7 +71,7 @@ impl EndpointIn {
         }
     }
 
-    pub fn configure(&self, _cs: &CriticalSection) {
+    pub fn configure(&self, _cs: CriticalSection<'_>) {
         if self.index() == 0 {
             let mpsiz = match self.descriptor.max_packet_size {
                 8 => 0b11,
@@ -97,7 +97,7 @@ impl EndpointIn {
         }
     }
 
-    pub fn deconfigure(&self, _cs: &CriticalSection) {
+    pub fn deconfigure(&self, _cs: CriticalSection<'_>) {
         let regs = self.usb.endpoint_in(self.index() as usize);
 
         // deactivating endpoint
@@ -163,7 +163,7 @@ impl EndpointOut {
         }
     }
 
-    pub fn configure(&self, _cs: &CriticalSection) {
+    pub fn configure(&self, _cs: CriticalSection<'_>) {
         if self.index() == 0 {
             let mpsiz = match self.descriptor.max_packet_size {
                 8 => 0b11,
@@ -189,7 +189,7 @@ impl EndpointOut {
         }
     }
 
-    pub fn deconfigure(&self, _cs: &CriticalSection) {
+    pub fn deconfigure(&self, _cs: CriticalSection<'_>) {
         let regs = self.usb.endpoint_out(self.index() as usize);
 
         // deactivating endpoint
@@ -205,11 +205,11 @@ impl EndpointOut {
     }
 
     pub fn read(&self, buf: &mut [u8]) -> Result<usize> {
-        interrupt::free(|cs| self.buffer.borrow(cs).borrow_mut().read_packet(buf))
+        critical_section::with(|cs| self.buffer.borrow_ref_mut(cs).read_packet(buf))
     }
 
     pub fn buffer_state(&self) -> EndpointBufferState {
-        interrupt::free(|cs| self.buffer.borrow(cs).borrow().state())
+        critical_section::with(|cs| self.buffer.borrow_ref(cs).state())
     }
 }